clear led on track stop

+++ api/src/led-matrix.ts
@@ -9,11 +9,17 @@ import {
 
 export default class LedMatrix {
   private childProcess: ChildProcess | null = null;
+  private previousCover: string | null = null;
 
   public async displayCover(filepath: string) {
+    if (this.previousCover === filepath) {
+      return;
+    }
+
     if (this.childProcess) {
       this.childProcess.kill();
     }
+    this.previousCover = filepath;
 
     this.childProcess = spawn(IMAGE_VIEWER_BINARY_PATH, [
       `--led-gpio-mapping=${GPIO_MAPPING}`,
@@ -21,5 +27,14 @@ export default class LedMatrix {
       `--led-brightness=${BRIGHTNESS}`,
       filepath,
     ]);
+  }
+
+  public clear() {
+    if (this.childProcess) {
+      this.childProcess.kill();
+    }
+
+    this.childProcess = null;
+    this.previousCover = null;
   }
 }
